backendai/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

# Middleware
python-multipart==0.0.9
aiofiles==23.2.1

# CORS
fastapi[all]

# Auth
PyJWT==2.8.0

# Env
python-dotenv==1.0.1

# OpenAI & Gemini
openai==1.58.1
google-generativeai

# MongoDB
pymongo==4.7.2

# AWS
boto3==1.34.118

# File Processing
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2

# Redis
redis==5.2.0

# HTML Parsing
beautifulsoup4==4.12.3

# Mail (optional - consider `fastapi-mail` instead of Flask-Mail)
# fastapi-mail==1.4.1

# Optional UI
gradio==4.27.0

# PostgreSQL
psycopg2-binary==2.9.10
